Se agrega Handlebars y Websockets, se agregan mejoras de visualización con CSS

+++ src/routes/products.routes.js
@@ -1,5 +1,6 @@
 import express from 'express';
-import { ProductManager } from '../ProductManager.js'
+import { ProductManager } from '../ProductManager.js';
+import { socketServer } from '../app.js';
 
 const router = express.Router();
 const productManager = new ProductManager('./products.json');
@@ -40,11 +41,14 @@ router.get('/:pid', async (req, res) => {
 
 router.post('/', async (req, res) => {
     try {
-        const { title, description, price, code, stock, category, thumbnails, status } = req.body;
+        const { title, description, price, code, stock, category, thumbnails } = req.body;
+        const status = (req.body.status === 'true');
         const isAdded = await productManager.addProduct(title, description, price, code, stock, category, thumbnails, status);
 
         if (isAdded) {
             res.send({ status: 'success', message: 'Product added successfully' });
+            const updatedProducts = await productManager.getProducts();
+            socketServer.emit('productAdded', updatedProducts);
         } else {
             res.status(400).send({ status: 'error', message: 'Failed to add product, code is repeated' });
         }
@@ -71,13 +75,19 @@ router.delete('/:pid', async (req, res) => {
     try {
 
         const id = parseInt(req.params.pid);
+        const product = await productManager.getProductById(id);
+        if (!product) {
+            return res.status(404).send({ error: 'Product does not exist' });
+        }
         await productManager.deleteProduct(id);
-
+        const updatedProducts = await productManager.getProducts();
+        socketServer.emit('productDeleted', { status: 'success', message: 'Product deleted successfully', updatedProducts });
         res.send({ status: 'success', message: 'Product deleted successfully' })
-
     } catch (error) {
         res.status(400).send({ status: 'error', error: 'Invalid request' });
     }
 })
+
+
 
 export default router;
